feat(eval): add provider config contract
