+++ .env.example
@@ -23,6 +23,17 @@ LANGFUSE_PUBLIC_KEY=
 LANGFUSE_SECRET_KEY=
 LANGFUSE_BASEURL=https://us.cloud.langfuse.com
 
+# Optional eval runner defaults. CLI flags override these values.
+# SQUIRE_EVAL_PROVIDER=anthropic
+# SQUIRE_EVAL_MODEL=claude-sonnet-4-6
+# SQUIRE_EVAL_RUN_LABEL=local-eval
+#
+# Future provider runner knobs. Leave unset until the provider runner supports them.
+# SQUIRE_EVAL_REASONING_EFFORT=
+# SQUIRE_EVAL_MAX_OUTPUT_TOKENS=
+# SQUIRE_EVAL_TIMEOUT_MS=
+# SQUIRE_EVAL_TOOL_LOOP_LIMIT=
+
 # Server environment — development enables dev-only routes and no-cache asset serving
 NODE_ENV=development
